feat: final audit & finishing - backend, desktop, mobile app
Backend:
- Added full REST API (auth, products, categories, stocks, shifts,
  transactions, devices, sync, reports, audit logs)
- Laravel Sanctum authentication with RBAC (admin/owner/cashier)
- CORS config for mobile web at localhost:8080
- Shift auto-creation fallback for offline transactions
- Idempotent transaction creation with UUID deduplication
- Stock adjustment accepts both quantity and quantity_change
- Device auto-registration in sync endpoints
- Expanded dashboard & report endpoints (items_sold, daily/weekly sales)
- Auth 401 returns JSON (not redirect)
- 25 feature tests passing (98 assertions)

Desktop App (Tauri + React):
- Fixed sync.ts: device_uuid support, sell_price/minimum_stock mapping
- Fixed CheckoutModal: dynamic shift ID from useShiftStore
- Added browser-compatible SQLite mock for web preview
- Added icon components (XIcon, BanknoteIcon, BuildingIcon, SmartphoneIcon)
- CheckoutModal UI polish: better icons, cleaner layout

Mobile App (Flutter):
- Fixed 15 compile/analyzer errors across core, models, features
- AppCache: web in-memory fallback (no SQLite FFI on web)
- AppConfig: auto-detect localhost vs Android emulator URL
- auth_repository: graceful error handling in restore()
- product_controller: fixed Future.wait usage
- main.dart: bootstrap() called on init
- flutter analyze: No issues found
- flutter test: All tests passed
- flutter build web: Built successfully

Screenshot:
- Added screenshot/desktop/ with 10 desktop menu screenshots
- Added screenshot/mobile/ directory

+++ mobile/lib/core/widgets/sales_bar_chart.dart
@@ -19,8 +19,9 @@ class SalesBarChart extends StatelessWidget {
 
     final visible =
         points.length > 14 ? points.sublist(points.length - 14) : points;
-    final maxValue = visible.map((p) => p.value).reduce((a, b) => a > b ? a : b);
-    final idealMax = (maxValue * 1.2).clamp(1000, double.infinity);
+    final maxValue =
+        visible.map((p) => p.value.toDouble()).reduce((a, b) => a > b ? a : b);
+    final idealMax = (maxValue * 1.2).clamp(1000, double.infinity).toDouble();
 
     return SizedBox(
       height: 180,
